Core: Separate markdown translation and path rewriting
