+++ src/co_op_translator/core/project/translation/__init__.py
@@ -1,5 +1,5 @@
 """Focused pieces that compose the project translation manager."""
 
-from .manager import TranslationManager
+from co_op_translator.core.project.translation.manager import TranslationManager
 
 __all__ = ["TranslationManager"]
