--- src/co_op_translator/utils/llm/markdown_utils.py
@@ -1,60 +1,0 @@
-"""Backward-compatible import path for markdown utilities.
-
-Implementation lives in :mod:`co_op_translator.utils.markdown`.
-"""
-
-from co_op_translator.utils.markdown import (
-    SPLIT_DELIMITER,
-    _group_lines_preserving_list_items,
-    _parse_markdown_text_and_code_parts,
-    _read_language_prompt_template,
-    build_translated_image_link,
-    compare_line_breaks,
-    count_links_in_markdown,
-    count_tokens,
-    extract_json_from_markdown_codeblock,
-    generate_evaluation_prompt,
-    generate_prompt_template,
-    get_tokenizer,
-    get_translated_markdown_dir,
-    migrate_notebook_links,
-    normalize_cjk_emphasis_markers,
-    normalize_internal_anchor_links,
-    process_markdown,
-    process_markdown_with_many_links,
-    replace_code_blocks,
-    restore_code_blocks,
-    split_markdown_content,
-    update_image_links,
-    update_links,
-    update_notebook_links,
-    update_untranslated_file_links,
-)
-
-__all__ = [
-    "SPLIT_DELIMITER",
-    "_group_lines_preserving_list_items",
-    "_parse_markdown_text_and_code_parts",
-    "_read_language_prompt_template",
-    "build_translated_image_link",
-    "compare_line_breaks",
-    "count_links_in_markdown",
-    "count_tokens",
-    "extract_json_from_markdown_codeblock",
-    "generate_evaluation_prompt",
-    "generate_prompt_template",
-    "get_tokenizer",
-    "get_translated_markdown_dir",
-    "migrate_notebook_links",
-    "normalize_cjk_emphasis_markers",
-    "normalize_internal_anchor_links",
-    "process_markdown",
-    "process_markdown_with_many_links",
-    "replace_code_blocks",
-    "restore_code_blocks",
-    "split_markdown_content",
-    "update_image_links",
-    "update_links",
-    "update_notebook_links",
-    "update_untranslated_file_links",
-]
